Final touches
I could not reproduce perfectly the truth directions found in the original code. I attributed this to the fact that truth directions were computed on a random sample of the training data while I use all the training data in my computation.

The truth direction I found is pretty close to the one original code found (cos = 0.9659)  but the polarity vector is not as close (cos=0.7891). Maybe this is due to random sampling maybe not. In both cases, I use all the training data to minimize variance.